refactor(github-ingest): restructure directory for blob and PR ingestion separation
Move blob-specific files to dedicated subdirectory to prepare for PR
ingestion: - Create github/ingest/blobs/ and
github/ingest/pull-requests/ directories - Move ingest-github-blobs.ts
and chunk-store.ts to blobs subdirectory - Update import paths to
reflect new structure - Keep common utilities (build-octokit, telemetry,
etc.) in parent directory

This change maintains backward compatibility while setting up the
structure for future PR ingestion implementation.
